fix(windows): mic audio control access and rename vibepollo_vmic
Expose audio_control_t helpers needed by the mic backend factory and rename apollo_vmic to vibepollo_vmic for upstream branding.
