+++ cmake/compile_definitions/windows.cmake
@@ -180,8 +180,8 @@ set(PLATFORM_TARGET_FILES
         "${CMAKE_SOURCE_DIR}/src/platform/windows/display_vram.cpp"
         "${CMAKE_SOURCE_DIR}/src/platform/windows/display_wgc.cpp"
         "${CMAKE_SOURCE_DIR}/src/platform/windows/audio.cpp"
-        "${CMAKE_SOURCE_DIR}/src/platform/windows/apollo_vmic.h"
-        "${CMAKE_SOURCE_DIR}/src/platform/windows/apollo_vmic.cpp"
+        "${CMAKE_SOURCE_DIR}/src/platform/windows/vibepollo_vmic.h"
+        "${CMAKE_SOURCE_DIR}/src/platform/windows/vibepollo_vmic.cpp"
         "${CMAKE_SOURCE_DIR}/src/platform/windows/mic_write.h"
         "${CMAKE_SOURCE_DIR}/src/platform/windows/mic_write.cpp"
         "${CMAKE_SOURCE_DIR}/src/platform/windows/virtual_display.h"
